Cleanup frontend code
- remove ng-bootstrap
- properly mock out services in specs
- use Orbit.uuid() to generate unique IDs
- remove utility.service.ts
- move notice component to xforge-common.module
- fix formatting in various files
- properly export components from xforge-common.module
