+++ src/SIL.XForge.Scripture/ClientApp/src/app/app.component.spec.ts
@@ -1,24 +1,24 @@
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { FormsModule } from '@angular/forms';
-import { MatProgressBarModule } from '@angular/material';
 import { RouterTestingModule } from '@angular/router/testing';
 import { OAuthService } from 'angular-oauth2-oidc';
 import { EMPTY } from 'rxjs';
 import { instance, mock, when } from 'ts-mockito';
 
 import { JSONAPIService } from '@xforge-common/jsonapi.service';
+import { XForgeCommonModule } from '@xforge-common/xforge-common.module';
 import { AppComponent } from './app.component';
+import { CoreModule } from './core/core.module';
 import { CounterComponent } from './counter/counter.component';
 import { FetchDataComponent } from './fetch-data/fetch-data.component';
 import { HomeComponent } from './home/home.component';
 import { NavMenuComponent } from './nav-menu/nav-menu.component';
-import { NoticeComponent } from './notice/notice.component';
 
 describe('AppComponent', () => {
-  const oauthServiceStub = mock(OAuthService);
-  const jsonApiServiceStub = mock(JSONAPIService);
-  when(oauthServiceStub.events).thenReturn(EMPTY);
+  const mockedOAuthService = mock(OAuthService);
+  const mockedJSONAPIService = mock(JSONAPIService);
+  when(mockedOAuthService.events).thenReturn(EMPTY);
 
   let fixture: ComponentFixture<AppComponent>;
 
@@ -29,22 +29,22 @@ describe('AppComponent', () => {
         NavMenuComponent,
         HomeComponent,
         CounterComponent,
-        FetchDataComponent,
-        NoticeComponent
+        FetchDataComponent
       ],
       imports: [
-        HttpClientTestingModule,
+        CoreModule,
         FormsModule,
-        MatProgressBarModule,
+        HttpClientTestingModule,
         RouterTestingModule.withRoutes([
           { path: 'home', component: HomeComponent },
           { path: 'counter', component: CounterComponent },
           { path: 'fetch-data', component: FetchDataComponent },
         ]),
+        XForgeCommonModule
       ],
       providers: [
-        { provide: OAuthService, useFactory: () => instance(oauthServiceStub) },
-        { provide: JSONAPIService, useFactory: () => instance(jsonApiServiceStub) }
+        { provide: OAuthService, useFactory: () => instance(mockedOAuthService) },
+        { provide: JSONAPIService, useFactory: () => instance(mockedJSONAPIService) }
       ]
     }).compileComponents();
   }));
