+++ src/SIL.XForge.Scripture/ClientApp/src/app/app.module.ts
@@ -17,10 +17,10 @@ import {
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { RouterModule } from '@angular/router';
-import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
+
+import { ChangePasswordComponent } from '@xforge-common/change-password/change-password.component';
+import { DetailSnackBarComponent } from '@xforge-common/notice/detail-snack-bar.component';
 import { XForgeCommonModule } from '@xforge-common/xforge-common.module';
-import { PasswordStrengthMeterModule } from 'angular-password-strength-meter';
-import { ChangePasswordComponent } from '../xforge-common/change-password/change-password.component';
 import { AppComponent } from './app.component';
 import { ConnectProjectComponent } from './connect-project/connect-project.component';
 import { CoreModule } from './core/core.module';
@@ -28,9 +28,6 @@ import { CounterComponent } from './counter/counter.component';
 import { FetchDataComponent } from './fetch-data/fetch-data.component';
 import { HomeComponent } from './home/home.component';
 import { NavMenuComponent } from './nav-menu/nav-menu.component';
-import { DetailSnackBarComponent } from './notice/detail-snack-bar.component';
-import { NoticeComponent } from './notice/notice.component';
-
 
 @NgModule({
   declarations: [
@@ -39,9 +36,6 @@ import { NoticeComponent } from './notice/notice.component';
     HomeComponent,
     CounterComponent,
     FetchDataComponent,
-    NoticeComponent,
-    DetailSnackBarComponent,
-    ChangePasswordComponent,
     ConnectProjectComponent
   ],
   imports: [
@@ -61,9 +55,6 @@ import { NoticeComponent } from './notice/notice.component';
     MatProgressSpinnerModule,
     MatSelectModule,
     MatSnackBarModule,
-    NgbModule,
-    PasswordStrengthMeterModule,
-    ReactiveFormsModule,
     ReactiveFormsModule,
     RouterModule.forRoot([
       { path: 'home', component: HomeComponent },
@@ -75,7 +66,7 @@ import { NoticeComponent } from './notice/notice.component';
     XForgeCommonModule
   ],
   providers: [],
-  entryComponents: [ DetailSnackBarComponent ],
+  entryComponents: [DetailSnackBarComponent],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
