--- src/SIL.XForge.Scripture/ClientApp/src/app/core/utility.service.ts
@@ -1,15 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable({
-  providedIn: 'root'
-})
-export class UtilityService {
-
-  static uuid(): string {
-    return 'xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx'.replace(/[x]/g, char => {
-      // tslint:disable-next-line:no-bitwise
-      const v = Math.random() * 16 | 0;
-      return v.toString(16);
-    });
-  }
-}
