--- src/SIL.XForge.Scripture/ClientApp/src/app/notice/notice.component.html
@@ -1,5 +1,0 @@
-<div *ngIf="isLoading()">
-    <div *ngIf="showProgressBar()">
-        <mat-progress-bar color="warn" [value]="getPercentComplete()"></mat-progress-bar>
-    </div>
-</div>
