--- src/SIL.XForge.Scripture/ClientApp/src/app/notice/notice.component.scss
@@ -1,3 +1,0 @@
-pre {
-    color: yellow;
-}
