+++ src/SIL.XForge.Scripture/ClientApp/src/xforge-common/change-password/change-password.component.html
@@ -3,17 +3,17 @@
     <mat-card-title class="changepassword-title">Change Password</mat-card-title>
     <mat-card-content fxLayout="column">
       <div fxLayout="column" fxLayoutAlign="center center">
-       <form [formGroup]="changePasswordForm" novalidate autocomplete="off" #changePasswordRef="ngForm">
-        <mat-form-field class="example-full-width">
-          <input matInput placeholder="New Password" type="password" id="newPassword" name="newPassword"
-            formControlName="newPassword" />
-          <mat-error *ngIf="formControls.newPassword.errors?.minlength">Password must be at least 7
-            characters</mat-error>
-          <mat-error *ngIf="formControls.newPassword.errors?.required && isSubmitted">Password is
-            required</mat-error>
-        </mat-form-field>
-        <password-strength-meter [password]="changePasswordForm.value.newPassword" item-width="200px"></password-strength-meter>
-        <mat-form-field class="example-full-width">
+        <form [formGroup]="changePasswordForm" novalidate autocomplete="off" #changePasswordRef="ngForm">
+          <mat-form-field class="example-full-width">
+            <input matInput placeholder="New Password" type="password" id="newPassword" name="newPassword"
+              formControlName="newPassword" />
+            <mat-error *ngIf="formControls.newPassword.errors?.minlength">Password must be at least 7
+              characters</mat-error>
+            <mat-error *ngIf="formControls.newPassword.errors?.required && isSubmitted">Password is
+              required</mat-error>
+          </mat-form-field>
+          <password-strength-meter [password]="changePasswordForm.value.newPassword" item-width="200px"></password-strength-meter>
+          <mat-form-field class="example-full-width">
             <input matInput placeholder="Confirm Password" type="password" id="confirmPassword" name="confirmPassword"
               formControlName="confirmPassword" />
             <mat-error *ngIf="formControls.confirmPassword.errors?.minlength">Confirm Password must be
@@ -25,11 +25,11 @@
             Passwords do not match
           </mat-error>
           <div class="changepassword-button">
-              <button mat-raised-button color="primary" id="btnChangePassword" (click)="onSubmit()"
-              type="submit" name="btnChangePassword" class="submit-button">Change Password</button>
+            <button mat-raised-button color="primary" id="btnChangePassword" (click)="onSubmit()" type="submit" name="btnChangePassword"
+              class="submit-button">Change Password</button>
           </div>
-          </form>
-        </div>
+        </form>
+      </div>
     </mat-card-content>
   </mat-card>
 </div>
