+++ src/SIL.XForge.Scripture/ClientApp/src/xforge-common/notice/notice.component.html
@@ -1,0 +1,5 @@
+<div *ngIf="isLoading()">
+  <div *ngIf="showProgressBar()">
+    <mat-progress-bar color="warn" [value]="getPercentComplete()"></mat-progress-bar>
+  </div>
+</div>
