+++ src/SIL.XForge.Scripture/ClientApp/src/xforge-common/notice/notice.component.scss
@@ -1,0 +1,3 @@
+pre {
+  color: yellow;
+}
